Exception handling implemented

+++ src/main/java/com/jonatasmateus/dscommerce/services/ProductService.java
@@ -3,10 +3,15 @@ package com.jonatasmateus.dscommerce.services;
 import com.jonatasmateus.dscommerce.dto.ProductDTO;
 import com.jonatasmateus.dscommerce.entities.Product;
 import com.jonatasmateus.dscommerce.repositories.ProductRepository;
+import com.jonatasmateus.dscommerce.services.exceptions.DatabaseException;
+import com.jonatasmateus.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -23,7 +28,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -36,14 +42,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        passDtoToEntity(dto, entity);
-        return new ProductDTO(repository.save(entity));
+        try {
+            Product entity = repository.getReferenceById(id);
+            passDtoToEntity(dto, entity);
+            return new ProductDTO(repository.save(entity));
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void passDtoToEntity(ProductDTO dto, Product entity) {
